+++ sample/src/main/java/me/relex/circleindicator/sample/fragment/DefaultFragment.java
@@ -8,7 +8,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import me.relex.circleindicator.CircleIndicator;
-import me.relex.circleindicator.sample.DemoPagerAdapter;
+import me.relex.circleindicator.sample.SamplePagerAdapter;
 import me.relex.circleindicator.sample.R;
 
 public class DefaultFragment extends Fragment {
@@ -16,13 +16,13 @@ public class DefaultFragment extends Fragment {
     @Nullable @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
             @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_default_demo, container, false);
+        return inflater.inflate(R.layout.fragment_sample_default, container, false);
     }
 
     @Override public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         ViewPager viewpager = (ViewPager) view.findViewById(R.id.viewpager);
         CircleIndicator indicator = (CircleIndicator) view.findViewById(R.id.indicator);
-        viewpager.setAdapter(new DemoPagerAdapter());
+        viewpager.setAdapter(new SamplePagerAdapter());
         indicator.setViewPager(viewpager);
         viewpager.setCurrentItem(2);
     }