+++ sample/src/main/java/me/relex/circleindicator/sample/fragment/LoopViewPagerFragment.java
@@ -3,24 +3,24 @@ package me.relex.circleindicator.sample.fragment;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import com.imbryk.viewPager.LoopViewPager;
 import me.relex.circleindicator.CircleIndicator;
-import me.relex.circleindicator.sample.SamplePagerAdapter;
 import me.relex.circleindicator.sample.R;
+import me.relex.circleindicator.sample.SamplePagerAdapter;
 
-public class CustomAnimationFragment extends Fragment {
+public class LoopViewPagerFragment extends Fragment {
 
     @Nullable @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
             @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_sample_custom_animation, container, false);
+        return inflater.inflate(R.layout.fragment_sample_loop_viewpager, container, false);
     }
 
     @Override public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
-        ViewPager viewpager = (ViewPager) view.findViewById(R.id.viewpager);
+        LoopViewPager viewpager = (LoopViewPager) view.findViewById(R.id.viewpager);
         CircleIndicator indicator = (CircleIndicator) view.findViewById(R.id.indicator);
         viewpager.setAdapter(new SamplePagerAdapter());
         indicator.setViewPager(viewpager);